changed the color of the chat header and chat footer, and added the link for the card suggestion button
